+++ examples/react/src/i18n/formatters.ts
@@ -5,7 +5,6 @@ import { date } from 'typesafe-i18n/formatters'
 export const initFormatters: FormattersInitializer<Locales, Formatters> = (locale: Locales) => {
 	const formatters: Formatters = {
 		weekday: date(locale, { weekday: 'long' }),
-
 	}
 
 	return formatters