+++ src/formatters/index.ts
@@ -1,4 +1,6 @@
 export { date, time } from './date'
+export { identity } from './identity'
+export { ignore } from './ignore'
 export { lowercase } from './lowercase'
 export { number } from './number'
 export { replace } from './replace'